Made only the last window closed save
This solution is non-ideal. I should come up with a mechanism
to save the state of all the windows upon application exist
instead of doing it on window close - that would allow an
environment like OS X to store the state for all open windows
while in Windows people would be greeted by the more familiar
window-per-process behavior.